requirements.txt
Flask
Gunicorn
requests
pymupdf
python-dotenv
bs4
